ISTE_Website_Django-master/requirements.txt
asgiref==3.2.7
certifi==2020.6.20
chardet==3.0.4
Django==3.0.5
django-crispy-forms==1.9.0
django-hosts==4.0
django-nucleus==0.1.2
django-rest-framework==0.1.0
djangorestframework==3.11.0
idna==2.10
pytz==2019.3
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
